fix: no hard fail on ThreadsafeFunction::call_async
For many reasons, in an app we can end up in a scenario where the thread / task
that initially called the ThreadsafeFunction gets interrupted. This means that
the receiver would get collected, should a ThreadsafeFunction be awaited on the
said thread. This will create an error when the ThreadsafeFunction's callback
will be called, and call a napi_fatal_error.

This change makes it so that if the send errors, this error is hidden and thus
prevents any hard failure. If a ThreadsafeFunction is called in such a case,
its output won't be used, but we'll still run all the logic to ensure it ran
properly.