fix(ThreadsafeFunction): ensure CalleeHandled works as expected
This was refactored in 4719caa64377f7d926c92a4c1051474ae79036c4. This codepath
was swapped, as the fatal exception must be the normal codepath when the callee
handled field is false. This fixes it by swapping this if statement, all the
other calls have been checked and seem fine.