Fix bigInt comparison
BigInt can representation can have trailing zeros,
and still represent the same logical value.

This comment replaces the derived implementation of == operator,
added in #3033, that didn't handle such cases correctly,
with manual implementation that allows for trailing zeros,
so that == return true, when BigInt represent the same logical value,
instead of the same in memory format.